requirements.txt
Django==1.9
django-aggregate-if==0.5
six==1.13.0
